+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - `cli/help`: Update help message
 - `cli/plugin`: Fix plugin options parser
+- `cli/cli`: Fix config options
 
 
 ## [0.0.14] - (2025-05-24)